Move the merge helpers to the stories that use them
Deleting the old merge test left the merge preview, apply, and two-paid-
duplicates set-up in the shared money drivers with only a story reaching
them, so the coverage gate failed — CI caught it.

They now live with the merge story, which also drops the second copy of the
preview I had written beside them: one reader opens the merge page and
returns the choices it offers, used both to answer them and to check no
money question is asked when nothing was paid.

The refund stand-in keeps its per-payment rule, which the bulk-refund story
needs, and now has direct tests for both forms.
